# backend/openapi.yaml
openapi: 3.1.0
info:
  title: RankingForge API
  version: 0.1.0
  description: >
    REST API for RankingForge. Authentication is via HttpOnly session cookies
    set on login. All endpoints except /auth/* require a valid session.
servers:
  - url: http://localhost:3000
components:
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
    User:
      type: object
      required: [id, username, created_at]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        created_at:
          type: string
          format: date-time
...
    Player:
      type: object
      required: [id, project_id, name, created_at, accounts]
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        name:
          type: string
        created_at:
          type: string
          format: date-time
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/StartggAccount'
    StartggAccount:
      type: object
      required: [id, startgg_user_id, slug]
      properties:
        id:
          type: string
          format: uuid
        startgg_user_id:
          type: integer
        slug:
          type: string
        display_name:
          type: string
          nullable: true
...
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnprocessableEntity:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
paths:
...
  /auth/register:
    post:
      summary: Create a new user account
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  minLength: 3
                password:
                  type: string
                  minLength: 8
      responses:
        '201':
          description: Account created; session cookie set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /auth/login:
    post:
      summary: Log in and receive a session cookie
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful; session cookie set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /projects/{project_id}/players:
    parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      summary: List all players in a project
      tags: [Players]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Player'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      summary: Add a player to the project
      tags: [Players]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{project_id}/players/{player_id}:
    parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: player_id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    delete:
      summary: Remove a player from the project
      tags: [Players]
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{project_id}/players/{player_id}/accounts:
    parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: player_id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    post:
      summary: Link a start.gg account to a player
      tags: [Players]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slug]
              properties:
                slug:
                  type: string
                  description: start.gg user slug, e.g. "user/abc123"
      responses:
        '201':
          description: Account linked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StartggAccount'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /projects/{project_id}/players/{player_id}/accounts/{account_id}:
    parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: player_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: account_id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    delete:
      summary: Unlink a start.gg account from a player
      tags: [Players]
      responses:
        '204':
          description: Unlinked
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Import ──────────────────────────────────────────────────────────────────
...